+++ addons/arsenal/config.cpp
@@ -19,7 +19,6 @@ class CfgPatches {
 #include "Display3DEN.hpp"
 #include "Cfg3DEN.hpp"
 #include "CfgEventHandlers.hpp"
-#include "CfgGestures.hpp"
 #include "CfgWeapons.hpp"
 #include "RscDisplayMain.hpp"
 #include "ACE_Arsenal_Sorts.hpp"
