+++ addons/arsenal/functions/fnc_showItem.sqf
@@ -58,7 +58,7 @@ if (_nextAction != GVAR(currentAction)) then {
     };
 
     if (simulationEnabled GVAR(center)) then {
-        GVAR(center) playActionNow QEGVAR(common,stop);
+        GVAR(center) playActionNow QEGVAR(common,stop); // resets gesture state (if arsenal is opened on animation transition, animations played whilst in the arsenal break)
         GVAR(center) playActionNow _nextAction;
     } else {
         GVAR(center) switchAction _nextAction;
